% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fBART-package.R
\docType{package}
\name{fBART-package}
\alias{fBART}
\alias{fBART-package}
\title{fBART: Basis-Expanded Bayesian Additive Regression Trees}
\description{
Fit Bayesian Additive Regreesion Trees using basis-expanded functional covariates.
}
\author{
\strong{Maintainer}: Joshua Marvald \email{joshua_marvald@urmc.rochester.edu}

}
\keyword{internal}
