% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BBARTselect.R
\name{fBARTselect}
\alias{fBARTselect}
\title{BBART variable selection}
\usage{
fBARTselect(
  y,
  x = NULL,
  func_data_list,
  num_trees = 100,
  num_samps = 5000,
  num_burn = 5000,
  num_thin = 5,
  num_chains = min(4, dbarts::guessNumCores()),
  num_threads_bart = min(num_chains, dbarts::guessNumCores()),
  num_threads_wrangle = dbarts::guessNumCores(),
  num_null_run = 10,
  prior_power = 4,
  prior_base = 0.95,
  time_points,
  break_spacing = "quantiles",
  n_breaks = 8,
  lambda_vals = 10^seq(-6, 6, length.out = 1000),
  min_method = "mean",
  min_quantile = NULL,
  manual_breaks = NULL,
  alpha_g = NULL,
  alpha_l = NULL,
  alpha_c = NULL
)
}
\arguments{
\item{y}{A vector of outcome values.}

\item{x}{A dataframe of scalar covariates. If 'nxp' then 'y' must have length 'n' and each dataframe in 'func_data_list' must have 'n' columns.}

\item{func_data_list}{List of functional covariates. Each list element should be a dataframe with rows indicating time points and columns for each subject. Importantly, each element of the list of functional covariates must be named.}

\item{num_trees}{Number of BART trees per sample.}

\item{num_samps}{Number of posterior samples to take after burn-in.}

\item{num_burn}{Number of burn-in samples.}

\item{num_thin}{Keep every 'num_thin' draw.}

\item{num_chains}{Number of independent MCMC chains to run.}

\item{num_threads_bart}{Number of threads used to run BART - not recommended to be greater than 'num_chains'.}

\item{num_threads_wrangle}{Number of threads for post processing - can be larger than 'num_chains'.}

\item{prior_power}{Power parameter for tree prior.}

\item{prior_base}{Base parameter for tree prior.}

\item{time_points}{Vector of observed time points. Method assumes all subjects and functional covariates are observed at a common set of time points.}

\item{break_spacing}{A value in c("quantiles", "equal", "manual") to set break spacing for B-spline basis.}

\item{n_breaks}{A number indicating number of breaks to use for B-spline basis (includes endpoints, results in n_breaks + 2 B-spline polynomials).}

\item{lambda_vals}{A vector of positive numbers to cross-validate over to get roughness penalty.}

\item{min_method}{A value to minimize during cross-validation, either "mean" or "quantile" (over the subjects GCV values)}

\item{min_quantile}{if min_method = "quantile", a value for quantile to use}

\item{manual_breaks}{An optional vector of manual break values. Must include domain endpoints.}

\item{alpha_g}{Determines the 1 - alpha_g quantile for a global threshold}

\item{alpha_l}{Determines the 1 - alpha_l quantile for separate variable thresholds}

\item{alpha_c}{Determines the 1 - alpha_c quantile for the global multiplier method}
}
\value{
A list of returns including any the appended data matrix, the TIPs from the real BBART run,
the basis used for the functional covariate basis expansions, any calculated thresholds, and any selected
scalar and functional covariates. Important domain regions can be identified by examining specific basis component
selections for functional covariates.
}
\description{
BBART variable selection
}
