% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_multiplier_thresh.R
\name{get_multiplier_thresh}
\alias{get_multiplier_thresh}
\title{Generate global multiplier thresholds}
\usage{
get_multiplier_thresh(record, alpha = 0.01)
}
\arguments{
\item{alpha}{Determines 1-alpha quantile in multiplier method}

\item{A}{A data frame of inclusion proportions ('p' columns and 'num_run' rows)}
}
\value{
A vector of variable specific global multiplier thresholds
}
\description{
Generate global multiplier thresholds
}
