% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_indicators.R
\name{make_indicators}
\alias{make_indicators}
\title{Function to return a vector of indicators}
\usage{
make_indicators(var_indices, p)
}
\arguments{
\item{var_indices}{A vector of integers specifying position of 1's in the output vector- largest value no greater than p.}

\item{p}{Length of output indicator vector}
}
\value{
A vector of 1's and 0's. 1's are in positions specified by var_indices.
}
\description{
Function to return a vector of indicators
}
\keyword{internal}
